pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lesion-tracker"
version = "0.1.0"
description = "A tool for tracking and labeling MRI lesions between baseline and follow-up scans"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Lesion Tracking Team"}
]
keywords = ["MRI", "lesion", "tracking", "multiple sclerosis", "neuroimaging", "medical imaging"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Scientific/Engineering :: Image Processing",
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "nibabel>=3.2.0",
    "scikit-image>=0.19.0",
    "matplotlib>=3.5.0",
    "pandas>=1.3.0",
    "click>=8.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.62.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
]
registration = [
    "antspyx>=0.3.0",
]
all = [
    "lesion-tracker[dev,registration]",
]

[project.scripts]
lesion-tracker = "lesion_tracker.cli:main"

[project.urls]
Homepage = "https://github.com/example/lesion-tracker"
Documentation = "https://github.com/example/lesion-tracker#readme"
Repository = "https://github.com/example/lesion-tracker"

[tool.setuptools.packages.find]
where = ["."]
include = ["lesion_tracker*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]

[tool.flake8]
max-line-length = 100
exclude = [".git", "__pycache__", "build", "dist"]
